Changes proposed in review of PR#8474
